+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
  * `Client#current_user` is a new function providing support for the `GET /api/whoami` endpoint.
  * `Client#auth_attempts_statistics` is a new function providing support for the `GET /api/auth/attempts/{node}` endpoint.
+ * `Client#list_topic_permissions` is a new function that provides support for the `GET /api/topic-permissions` endpoint.
+ * `Client#list_topic_permissions_in` is a new function that provides support for the `GET /api/vhosts/{vhost}/topic-permissions` endpoint.
+ * `Client#get_topic_permissions_of` is a new function that provides support for the `GET /api/topic-permissions/{vhost}/{user}` endpoint.
 
 
 ## v0.50.0 (Sep 13, 2025)